Add deprecated device warning
