+++ app/scripts/i18n/react/locales/ru/translations.json
@@ -242,7 +242,8 @@
       },
       "errors": {
          "write": "Ошибка при сохранении настроек. Более детальная информация находится в журнале работы сервиса wb-mqtt-confed",
-         "device-config": "Ошибка в настройках одного или нескольких устройств"
+         "device-config": "Ошибка в настройках одного или нескольких устройств",
+         "deprecated": "Шаблон устройства устарел, рекомендуется использовать новую версию"
       }
    }
 }
