+++ app/scripts/react-directives/device-manager/jsonSchemaUtils.js
@@ -77,3 +77,7 @@ export function getDefaultObject(schema, root, defaultProperties) {
   }
   return res;
 }
+
+export function getTranslation(key, lang, translations) {
+  return translations[lang]?.[key] || translations?.en?.[key] || key;
+}
